jit: (llvm) rename convert_ctypes to ctype_context
